refactor: move SessionSetup back to pages/chat as page-specific component
SessionSetup will differ between agent pages (chat vs coding), so it
should not be in the shared module.
