+++ apps/frontend/src/pages/chat/components/SessionSetup/components/ExtraAllowedPathsSelect/ExtraAllowedPathsSelectView.tsx
