+++ apps/frontend/src/pages/chat/components/SessionSetup/components/ExtraAllowedPathsSelect/index.ts
