+++ apps/frontend/src/pages/chat/components/SessionSetup/components/WorkspaceSelect/index.ts
