+++ apps/frontend/src/pages/chat/components/SessionSetup/components/WorkspaceSelect/styles.module.css
