+++ apps/frontend/src/pages/chat/components/SessionSetup/styles.module.css
